chore: update package name
