+++ jnosql-lite/mapping-lite-core-test/src/main/java/org/eclipse/jnosql/mapping/entities/Car.java
@@ -1,5 +1,5 @@
 /*
- *  Copyright (c) 2023 Otávio Santana and others
+ *  Copyright (c) 2020 Otávio Santana and others
  *   All rights reserved. This program and the accompanying materials
  *   are made available under the terms of the Eclipse Public License v1.0
  *   and Apache License v2.0 which accompanies this distribution.
@@ -12,60 +12,38 @@
  *
  *   Otavio Santana
  */
-package org.eclipse.jnosql.mapping.lite;
+package org.eclipse.jnosql.mapping.entities;
 
 import jakarta.nosql.Column;
 import jakarta.nosql.Entity;
 import jakarta.nosql.Id;
 
-import java.util.List;
-
-
-@Entity
-public class User {
+@Entity("car")
+public class Car {
 
     @Id
-    private long id;
-
-    @Column
     private String name;
 
     @Column
-    private int age;
+    private String model;
 
-    @Column
-    private List<String> phones;
-
-
-    public long getId() {
-        return id;
+    Car() {
     }
 
-    public void setId(long id) {
-        this.id = id;
-    }
 
-    public String getName() {
+    String getName() {
         return name;
     }
 
-    public void setName(String name) {
+    void setName(String name) {
         this.name = name;
     }
 
-    public int getAge() {
-        return age;
+    protected String getModel() {
+        return model;
     }
 
-    public void setAge(int age) {
-        this.age = age;
-    }
-
-    public List<String> getPhones() {
-        return phones;
-    }
-
-    public void setPhones(List<String> phones) {
-        this.phones = phones;
+    protected void setModel(String model) {
+        this.model = model;
     }
 }
