+++ jnosql-lite/mapping-lite-core-test/src/main/java/org/eclipse/jnosql/mapping/entities/Money.java
@@ -12,7 +12,7 @@
  *
  *   Otavio Santana
  */
-package org.eclipse.jnosql.mapping.lite;
+package org.eclipse.jnosql.mapping.entities;
 
 import java.math.BigDecimal;
 import java.util.Objects;
