+++ jnosql-lite/mapping-lite-core-test/src/main/java/org/eclipse/jnosql/mapping/entities/Orders.java
@@ -12,40 +12,38 @@
  *
  *   Otavio Santana
  */
-package org.eclipse.jnosql.mapping.lite;
-
+package org.eclipse.jnosql.mapping.entities;
 
 import jakarta.nosql.Column;
 import jakarta.nosql.Entity;
-import jakarta.nosql.Id;
 
+import java.util.List;
 import java.util.Objects;
 
-@Entity("kind")
-public class Animal {
-
-    @Id
-    private String name;
+@Entity
+public class Orders {
 
     @Column
-    private String color;
+    private String user;
 
-    public String getName() {
-        return name;
+    @Column
+    private List<Product> items;
+
+    public String getUser() {
+        return user;
     }
 
-    void setName(String name) {
-        this.name = name;
+    public void setUser(String user) {
+        this.user = user;
     }
 
-    public String getColor() {
-        return color;
+    public List<Product> getItems() {
+        return items;
     }
 
-    void setColor(String color) {
-        this.color = color;
+    public void setItems(List<Product> items) {
+        this.items = items;
     }
-
 
     @Override
     public boolean equals(Object o) {
@@ -55,21 +53,12 @@ public class Animal {
         if (o == null || getClass() != o.getClass()) {
             return false;
         }
-        Animal animal = (Animal) o;
-        return Objects.equals(name, animal.name) &&
-                Objects.equals(color, animal.color);
+        Orders that = (Orders) o;
+        return Objects.equals(user, that.user);
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(name, color);
-    }
-
-    @Override
-    public String toString() {
-        return "Animal{" +
-                "name='" + name + '\'' +
-                ", color='" + color + '\'' +
-                '}';
+        return Objects.hashCode(user);
     }
 }
