+++ jnosql-lite/mapping-lite-core-test/src/test/java/org/eclipse/jnosql/mapping/lite/EntitiesMetadataTest.java
@@ -14,6 +14,9 @@
  */
 package org.eclipse.jnosql.mapping.lite;
 
+import org.eclipse.jnosql.mapping.entities.Animal;
+import org.eclipse.jnosql.mapping.entities.Car;
+import org.eclipse.jnosql.mapping.entities.Person;
 import org.eclipse.jnosql.mapping.metadata.EntitiesMetadata;
 import org.eclipse.jnosql.lite.mapping.metadata.LiteEntitiesMetadata;
 import org.eclipse.jnosql.mapping.metadata.EntityMetadata;
