+++ jnosql-lite/mapping-lite-core-test/src/test/java/org/eclipse/jnosql/mapping/lite/MovieTest.java
@@ -15,6 +15,8 @@
 package org.eclipse.jnosql.mapping.lite;
 
 
+import org.eclipse.jnosql.mapping.entities.Director;
+import org.eclipse.jnosql.mapping.entities.Movie;
 import org.eclipse.jnosql.mapping.metadata.EntitiesMetadata;
 import org.eclipse.jnosql.lite.mapping.metadata.LiteEntitiesMetadata;
 import org.eclipse.jnosql.mapping.metadata.EntityMetadata;
