+++ jnosql-lite/mapping-lite-core-test/src/test/java/org/eclipse/jnosql/mapping/lite/inheritance/EmailNotificationTest.java
@@ -14,6 +14,8 @@
  */
 package org.eclipse.jnosql.mapping.lite.inheritance;
 
+import org.eclipse.jnosql.mapping.entities.inheritance.EmailNotification;
+import org.eclipse.jnosql.mapping.entities.inheritance.Notification;
 import org.eclipse.jnosql.mapping.metadata.EntitiesMetadata;
 import org.eclipse.jnosql.lite.mapping.metadata.LiteEntitiesMetadata;
 import org.eclipse.jnosql.mapping.metadata.EntityMetadata;
